+++ src/main/java/techcourse/myblog/dto/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain.dto;
+package techcourse.myblog.dto;
 
 public class LoginRequestDto {
     private String email;
