+++ src/main/java/techcourse/myblog/web/MainController.java
@@ -4,8 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import techcourse.myblog.domain.dto.UserRequestDto;
 import techcourse.myblog.domain.service.ArticleService;
+import techcourse.myblog.dto.UserRequestDto;
 
 import javax.servlet.http.HttpSession;
