+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -6,9 +6,9 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.domain.User;
-import techcourse.myblog.domain.dto.MyPageRequestDto;
-import techcourse.myblog.domain.dto.UserRequestDto;
 import techcourse.myblog.domain.service.UserService;
+import techcourse.myblog.dto.MyPageRequestDto;
+import techcourse.myblog.dto.UserRequestDto;
 
 import javax.servlet.http.HttpSession;
 import javax.transaction.Transactional;
